requirements.txt
altgraph==0.17.4
contourpy==1.2.1
cycler==0.12.1
fonttools==4.53.1
kiwisolver==1.4.5
matplotlib==3.9.1
numpy==2.0.0
packaging==24.1
pandas==2.2.2
pefile==2023.2.7
pillow==10.4.0
pyinstaller==6.9.0
pyinstaller-hooks-contrib==2024.7
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
pywin32-ctypes==0.2.2
setuptools==71.1.0
six==1.16.0
tenacity==8.5.0
tzdata==2024.1
